convert menu and tooltip to use mount / unmount
