+++ frontend/openchat-client/src/stores/iconSize.ts
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const iconSize = writable<string>("1.4em");
